LayoutTests/inspector/editor/text-editor-remove-chunks-from-dom-expected.txt

This test checks how damaged chunks' DOM is removed from the container.


Running: testDamageLastChunk

Before editing
Text model
[0]       0
[1]       1
[2]       2
[3]       3
[4]       4
[5]       
Editor DOM
[0]       <div class="webkit-line-content">0</div>
[1]       <div class="webkit-line-content">1</div>
[2]       <div class="webkit-line-content">2</div>
[3]       <div class="webkit-line-content">3</div>
[4]       <div class="webkit-line-content">4</div>
[5]       <div class="webkit-line-content"><br></div>

After editing
Text model
[0]       0
[1]       1
[2]       2
[3]       3
[4]       4
[5]       additional 5
[6]       
Editor DOM
[0]       <div class="webkit-line-content">0</div>
[1]       <div class="webkit-line-content">1</div>
[2]       <div class="webkit-line-content">2</div>
[3]       <div class="webkit-line-content">3</div>
[4]       <div class="webkit-line-content">4</div>
[5]       <div class="webkit-line-content">additional 5</div>
[6]       <div class="webkit-line-content"><br></div>
